refactor(api): extract /api/status and /api/v2/status into routes/status.py
